Input type number and validations on the addRecipe and confirm before delete a recipe

+++ src/Components/Admin/AdminRecipes/AdminRecipes.jsx
@@ -52,18 +52,21 @@ function Admin() {
 
   /*Delete*/
   const remove = (id) => {
+    let confirmation = window.confirm("Esta seguro de eliminar esta receta?");
     const deleteRecipe = recipes.find(recipe => recipe._id === id);
-    fetch(
-      `https://polar-reaches-30197.herokuapp.com/recipes/${deleteRecipe._id}`,
-      {
-        method: "DELETE",
-        body: JSON.stringify(),
-        headers: {
-          "Content-Type": "application/json",
-          "x-access-token": userToken,
-        },
-      }
-    ).then((response) => setFetchFlag(response.json()));
+    if(confirmation){
+      fetch(
+        `https://polar-reaches-30197.herokuapp.com/recipes/${deleteRecipe._id}`,
+        {
+          method: "DELETE",
+          body: JSON.stringify(),
+          headers: {
+            "Content-Type": "application/json",
+            "x-access-token": userToken,
+          },
+        }
+      ).then((response) => setFetchFlag(response.json()));
+    }    
   };
 
   return (
